Fixed bug in BT; cleanup
