+++ VM/Registers.py
@@ -1,244 +1,221 @@
-from .util import byteorder, to_int, SegmentRegs, segment_descriptor_struct
+import ctypes
+
+from .ctypes_types import ubyte, uword, udword
+
+__all__ = 'Reg32',
+
+REG_LETTERS = 'acdb'
+REG_TAILS = 'sp', 'bp', 'si', 'di'
 
 
-class SegmentReg:
-    def __init__(self, name: str):
-        self.segment_selector = 0
-        self.__name = name
-        self.__hidden = [0, 0, 0]  # base address, limit, access info
+def GenRegX(letter: str):
+    assert letter in REG_LETTERS
 
-        self.__bytes = None
+    class _HiLo_byte(ctypes.LittleEndianStructure):
+        _pack_ = 1
+        # these are SWAPPED because of endianness
+        _fields_ = [
+            (letter + 'l', ubyte),
+            (letter + 'h', ubyte),
+        ]
 
-    @property
-    def base(self):
-        return self.__hidden[0]
+    # ctypes' docs (16.12.1.11. Structure/union alignment and byte order) say LittleEndianUnion exists, like LittleEndianStructure but it doesn't
+    class _Reg32(ctypes.Union):
+        _pack_ = 1
+        _anonymous_ = '__byte',
+        _fields_ = [
+            (f'e{letter}x', udword),
+            (letter + 'x', uword),
+            ('__byte', _HiLo_byte)
+        ]
 
-    @property
-    def limit(self):
-        return self.__hidden[1]
+    return _Reg32
 
-    def from_bytes(self, seg_sel: int, cache: bytes):
-        self.segment_selector = seg_sel
 
-        base3, _limit2, info, base2, base1, limit1 = segment_descriptor_struct.unpack(cache)
+def GenReg(tail: str):
+    assert tail in REG_TAILS
 
-        limit2 = _limit2 & 0xF
-        limit = (limit2 << 16) + limit1
-        base = (base3 << (3 * 8)) + (base2 << (2 * 8)) + base1
+    class _Reg32(ctypes.Union):
+        _fields_ = [
+            (f'e{tail}', udword),
+            (tail, uword),
+        ]
 
-        self.__hidden[:2] = base, limit
-        # TODO: do  something with access data
+    return _Reg32
 
-        self.__bytes = None
 
-    def from_data(self, seg_sel: int, base: int, limit: int):
-        self.segment_selector = seg_sel
-        self.__hidden[:2] = base, limit
+class _Eflags_bits(ctypes.LittleEndianStructure):
+    _pack_ = 1
+    _fields_ = [
+        ('CF', uword, 1),
+        ('__', uword, 1),
+        ('PF', uword, 1),
+        ('__', uword, 1),
+        ('AF', uword, 1),
+        ('__', uword, 1),
+        ('ZF', uword, 1),
+        ('SF', uword, 1),
+        ('TF', uword, 1),
+        ('IF', uword, 1),
+        ('DF', uword, 1),
+        ('OF', uword, 1),
+    ][::-1]
 
-        # TODO: do  something with access data
 
-        self.__bytes = None
+class _Eflags(ctypes.Union):
+    _anonymous_ = '__bits',
+    _fields_ = [
+        ('eflags', uword),
+        ('__bits', _Eflags_bits)
+    ]
+
+
+class _Reg32_base(ctypes.LittleEndianStructure):
+    a = ['reg_' + l for l in REG_LETTERS]
+    b = ['reg_' + t for t in REG_TAILS]
+
+    _anonymous_ = tuple(a + b)
+
+    _fields_ = [
+                   (name, GenRegX(name[4:]))
+                   for name in a
+               ] + [
+                   (name, GenReg(name[4:]))
+                   for name in b
+               ] + [
+        ('eflags', _Eflags)
+    ][::-1]
+
+    del a, b
+
+
+class _sreg_hidden(ctypes.LittleEndianStructure):
+    _pack_ = 1
+    _fields_ = [
+        ('base', udword),
+        ('limit', udword),
+        ('access', udword)
+    ]
 
     def __str__(self):
-        return self.__name
-
-    def __bytes__(self):
-        if self.__bytes is None:
-            self.__bytes = self.segment_selector.to_bytes(2, byteorder) +\
-                           self.__hidden[0].to_bytes(4, byteorder) +\
-                           self.__hidden[1].to_bytes(3, byteorder) +\
-                           self.__hidden[2].to_bytes(1, byteorder)
-
-        return self.__bytes
+        return f'base=0x{self.base:08x}, limit=0x{self.limit:08x}, access={self.access:08x}'
 
 
-class Reg32:
+class _one_sreg(ctypes.Structure):
+    _pack_ = 1
+    _fields_ = [
+        ('visible', udword),
+        ('hidden', _sreg_hidden)
+    ]
+
+    def __str__(self):
+        return f'Sreg(visible=0x{self.visible:08x}, {self.hidden})'
+
+
+class SegmentDescriptor(ctypes.LittleEndianStructure):
+    _pack_ = 1
+    _fields_ = [
+        ('base_3', ubyte),
+        ('G', ubyte, 1),
+        ('DB', ubyte, 1),
+        ('L', ubyte, 1),
+        ('AVL', ubyte, 1),
+        ('limit_2', ubyte, 4),
+        ('P', ubyte, 1),
+        ('DPL', ubyte, 2),
+        ('S', ubyte, 1),
+        ('type', ubyte, 4),
+        ('base_2', ubyte),
+
+        ('base_1', uword),
+        ('limit_1', uword)
+    ][::-1]
+
+    def __str__(self):
+        return f'SegmentDescriptor(base3={self.base_3:08b}, G={self.G:1b}, DB={self.DB:1b}, L={self.L:1b}, \
+AVL={self.AVL:1b}, limit2={self.limit_2:04b}, P={self.P:1b}, DPL={self.DPL:02b}, S={self.S:1b}, \
+type={self.type:04b}, base2={self.base_2:08b}, base1={self.base_1:016b}, limit1={self.limit_1:016b})'
+
+
+class Sreg(ctypes.LittleEndianStructure):
     """
-    Provides all the 32-, 16- and 8-bit registers of a IA-32 CPU.
+    Segment registers.
     """
-    names = "eax ecx edx ebx esp ebp esi edi".split()
+    _pack_ = 1
+    _fields_ = [
+        (name, _one_sreg)
+        for name in 'ES CS SS DS FS GS'.split()
+    ]
 
-    CF, PF, AF, ZF, SF, TF, IF, DF, OF = 0, 2, 4, *range(6, 12)  # EFLAGS bits
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
-    def __init__(self):
-        self.allowed_sizes = [4, 2, 1]
-        # self.registers = bytearray(self.allowed_sizes[0] * len(Reg32.names))
-        self.eflags = 0x02  # initial value according to the Intel Software Development Manual
+        self.__names = 'ES CS SS DS FS GS'.split()
+        self.__ptr = ctypes.cast(ctypes.pointer(self), ctypes.POINTER(_one_sreg))
 
-        self.sreg = [SegmentReg(n) for n in SegmentRegs._member_names_] # segment registers
+    def set(self, offset: int, segment_selector: int, descriptor_data: bytes) -> None:
+        # index, TI, RPL = segment_selector >> 3, (segment_selector >> 2) & 1, segment_selector & 0b11
+        # print(f'offset={offset:03b}, (index={index:05b}, TI={TI:1b}, RPL={RPL:02b}), descriptor: {descriptor_data}')
 
-        self.GDTR = 0xFFFF  # Global Descriptor Table Register (see vol. 3A 2.4.1)
-        self.LDTR = 0xFFFF  # Local Descriptor Table Register (see vol. 3A 2.4.2)
+        descriptor = SegmentDescriptor.from_buffer_copy(descriptor_data)
 
-        self.bounds = range(len(Reg32.names))
+        # print(f'Descriptor: {descriptor}, descr. data=={descriptor_data}')
 
-        reserved_eflags_bits = {1, 3, 5, 15}
-        reserved_eflags_bits.update(set(range(22, 32)))
-        self.eflags_bounds = set(range(32)) - reserved_eflags_bits
+        # TODO: handle priviledge level
 
-        registers = [
-            b'\0\0\0\0',  # EAX (000)
-            b'\0\0\0\0',  # ECX (001)
-            b'\0\0\0\0',  # EDX (010)
-            b'\0\0\0\0',  # EBX (011)
-            b'\0\0\0\0',  # ESP (100)
-            b'\0\0\0\0',  # EBP (101)
-            b'\0\0\0\0',  # ESI (110)
-            b'\0\0\0\0',  # EDI (111)
-        ]
-        self.registers = list(map(bytearray, registers))
+        self.__ptr[offset].visible = segment_selector
+        self.__ptr[offset].hidden.base = (descriptor.base_3 << 24) | (descriptor.base_2 << 16) | descriptor.base_1
+        self.__ptr[offset].hidden.limit = (descriptor.limit_2 << 16) | descriptor.limit_1
 
-    def get(self, offset: int, size: int) -> bytes:
-        """
-        Extract `size` bytes and reverse them if byteorder == 'little'
-        :param offset: Register number according to the Intel Software Developer Manual
-        :param size: Size of the register to read, in bytes.
-        :return: The actual bytes read.
-        """
+        # print(f'segment.visible={self.__ptr[offset].visible}, segment.hidden.base=0x{self.__ptr[offset].hidden.base:08x}, segment.hidden.limit=0x{self.__ptr[offset].hidden.limit:08x}')
 
-        # print(f'Reg32.get({offset:03b}, {size}) -> ', end='')
-        if size == 4 or size == 2:
-            __tmp = self.registers[offset] #[::-1]
-            val = __tmp[:size]
-        elif size == 1:
-            __tmp = self.registers[offset & 0b011]  # get rid of the first 1 bit
-            val = bytes([__tmp[(offset >> 2) & 1]])
-        else:
-            raise RuntimeError(f"Invalid offset for PyVM.Registers.Reg32.get: {offset}")
+    def get(self, offset: int) -> _one_sreg:
+        assert 0b000 <= offset <= 0b101
 
-        # print(f'{__tmp.hex()} -> {val.hex()}')
+        return self.__ptr[offset]
 
-        return val
 
-    def set(self, offset: int, value: bytes) -> None:
-        """
-        Set literal value, reversed in little-endian!
-        :param offset:
-        :param value:
-        :return:
-        """
+class Reg32(_Reg32_base):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
-        size = len(value)
+        ptr = ctypes.pointer(self)
 
-        # print(f'Setting {value.hex()}')
+        self.__ptr8 = ctypes.cast(ptr, ctypes.POINTER(ubyte))
+        self.__ptr16 = ctypes.cast(ptr, ctypes.POINTER(uword))
+        self.__ptr32 = ctypes.cast(ptr, ctypes.POINTER(udword))
 
+    def get(self, offset: int, size: int, signed=False) -> int:
         if size == 4:
-            self.registers[offset][:] = value
+            if not signed:
+                return self.__ptr32[offset]
+
+            ret = self.__ptr32[offset]
+            return ret if ret < 2147483648 else ret - 4294967296
         elif size == 2:
-            self.registers[offset][:2] = value
+            if not signed:
+                return self.__ptr16[2 * offset]
+
+            ret = self.__ptr16[2 * offset]
+            return ret if ret < 32768 else ret - 65536
         elif size == 1:
-            self.registers[offset & 0b011][(offset >> 2) & 1] = value[0]
+            if not signed:
+                return self.__ptr8[4 * (offset % 4) + offset // 4]
+
+            ret = self.__ptr8[4 * (offset % 4) + offset // 4]
+            return ret if ret < 128 else ret - 256
+
+        raise ValueError(f'Reg32.get(offset={offset}, size={size}): unexpected size: {size}')
+
+    def set(self, offset: int, size: int, value: int) -> None:
+        if size == 4:
+            self.__ptr32[offset] = value
+        elif size == 2:
+            self.__ptr16[2 * offset] = value
+        elif size == 1:
+            self.__ptr8[4 * (offset % 4) + offset // 4] = value
         else:
-            raise RuntimeError(f"Invalid offset for PyVM.Registers.Reg32.set: {offset}")
+            raise ValueError(f'Reg32.set_val(offset={offset}, size={size}, value={value}): unexpected size: {size}')
 
-    def eflags_get(self, bit: int) -> int:
-        """
-        Retrieve a specific bit of the EFLAGS register
-        :param bit: the number of the bit
-        :return:
-        """
-        assert bit in self.eflags_bounds, 'Reg32.eflags_get: invalid bit number {} (allowed bit numbers: {})'.format(bit, self.eflags_bounds)
-
-        return (self.eflags >> bit) & 1
-
-    def eflags_set(self, bit: int, value: bool) -> None:
-        """
-        Set a specific bit of the EFLAGS register to some value
-        :param bit: the number of the bit
-        :param value: boolean
-        :return:  None
-        """
-        if self.eflags_get(bit):
-            if not value:
-                self.eflags &= ~(1 << bit)
-        else:
-            if value:
-                self.eflags |= 1 << bit
-
-
-class FReg32:
-    names = "st0 st1 st2 st3 st4 st5 st6 st7".split()
-
-    def __init__(self):
-        self.allowed_sizes = [8]
-        self.registers = bytearray(self.allowed_sizes[0] * len(FReg32.names))
-
-        self.control = 0
-        self.status = 0b00_111_00000000000
-        self.tag = 0xFFFF
-
-    def _get_st(self, reg_num: int) -> bytes:
-        TOP = self.TOP + reg_num
-
-        if TOP >= 0b111:
-            raise RuntimeError("The floating-point TOP pointer is messed up!")
-
-        sz = self.allowed_sizes[0]
-
-        s = slice(TOP * sz, (TOP + 1) * sz)
-        value = self.registers[TOP * sz:(TOP + 1) * sz]
-        #print(f"Actually getting ST{reg_num} ({s}) -> {value}")
-        return value
-
-    def _set_st(self, reg_num: int, value: bytes) -> None:
-        sz = self.allowed_sizes[0]
-        assert len(value) == sz
-
-        TOP = self.TOP + reg_num
-
-        if TOP < 0:
-            raise RuntimeError("The floating-point TOP pointer is messed up!")
-
-        s = slice(TOP * sz, (TOP + 1) * sz)
-
-        #print(f"Actually setting ST{reg_num} ({s}) = {value}")
-        self.registers[s] = value
-
-    # TODO: generate that with a metaclass
-    ST0 = property(lambda self: self._get_st(0), lambda self, val: self._set_st(0, val))
-    ST1 = property(lambda self: self._get_st(1), lambda self, val: self._set_st(1, val))
-    ST2 = property(lambda self: self._get_st(2), lambda self, val: self._set_st(2, val))
-    ST3 = property(lambda self: self._get_st(3), lambda self, val: self._set_st(3, val))
-    ST4 = property(lambda self: self._get_st(4), lambda self, val: self._set_st(4, val))
-    ST5 = property(lambda self: self._get_st(5), lambda self, val: self._set_st(5, val))
-    ST6 = property(lambda self: self._get_st(6), lambda self, val: self._set_st(6, val))
-    ST7 = property(lambda self: self._get_st(7), lambda self, val: self._set_st(7, val))
-
-    @property
-    def TOP(self) -> int:
-        return (self.status >> 11) & 0b111
-
-    @TOP.setter
-    def TOP(self, val: int) -> None:
-        self.status &= 0b11_000_11111111111  # clear TOP; ugly AF
-        self.status |= val << 11  # set TOP back
-
-    def push(self, value: bytes) -> None:
-        sz = self.allowed_sizes[0]
-
-        assert len(value) == sz
-
-        TOP = self.TOP
-
-        if TOP == 0:
-            raise ValueError("Push onto full stack")
-
-        # print(f"Pushing {value} -> {TOP * sz}:{(TOP + 1) * sz} ({TOP}:{TOP + 1})")
-
-        self.TOP = TOP - 1
-        self.ST0 = value
-
-    def pop(self, increase_TOP=True) -> bytearray:
-        sz = self.allowed_sizes[0]
-        TOP = self.TOP
-
-        if TOP == 0b111:
-            raise ValueError("Pop from empty stack")
-
-        # print(f"Popping {TOP * sz}:{(TOP + 1) * sz}")
-        ret = self.ST0
-
-        if increase_TOP:
-            self.TOP = TOP + 1
-
-        return ret
+    def set4(self, offset: int, value: int):
+        self.__ptr32[offset] = value
